feat: refactor LLM client layer and update provider identifiers
- Refactored the LLM client layer from a monolithic structure in `clients.py` to an adapter-based architecture in the new `src/utils/llm/` package, enhancing modularity and maintainability.
- Renamed the provider identifier from `"custom"` to `"openrouter"` for improved clarity in configuration settings.
- Updated relevant files to reflect the new structure and naming conventions, ensuring consistency across the codebase.
- Enhanced the changelog to document these significant changes.
